Feature/global ens priorities 456 plots (#348)
* Replace CYC with new, appropriate variables names

Refs: #5

* Replace cycles with new, appropriate variables names

Refs: #5

* Replace cycle with new, appropriate variables names

Refs: #5

* Replace cyc with new, appropriate variables names

Refs: #5

* Add vhr in J-Jobs

Refs: #5

* Add vhr in wave dev drivers scripts

Refs: #8

* Use ihour instead of vhour

Refs: #5

* Remove unused environment variables

Refs: #5

* Update script name

Refs: #5

* Add evs_ver and evs_ver_2d

Refs: #5

* Remove unused global_ens files in metplus_config

Refs: #5

* Remove more unused global_ens files in metplus_config

Refs: #5

* Remove MORE unused global_ens files in metplus_config

Refs: #5

* Add LOG_METPLUS as empty string

Refs: #5

* Cat the plotting log files

Refs: #5

* Add set -x

Refs: #5

* Remove unused ush scripts

Refs: #5

* Remove exevs_global_ens_aerosol_stats.sh

Refs: #5

* Add interpreter at the top of scripts

Refs: #5

* Remove METplus machine.conf settings

Refs: #5

* Remove hard-coded paths

Refs: #5

* List all possible missing CCPA files in missing email message

Refs: #5

* Fix wave plots legends

Refs: #5

* Remove jlogfile and postmsg usage

Refs: #5

* Add restrictions for rstprod data

Refs: #5

* Remove more hard-coded paths

Refs: #5

* Final EE2 change for global_ens wave

* Save atmos prep changes

* Remove vhr

* Clean up empty lines

* Use array to gather all avaiable fhrs to run METplus once

* Fix resources

* First round of editing error cases to use keywords and to handle the error properly

* Fix fatal error cases to follow standard procedure

* Add err_chk and err_exit statements + clarify error messages + add some commentary

* Change "cp" to "cpreq" where required

* Remove unecessary commented-out code

* check that origin file exists before mv command

* Fix indentation error in python + surround tar command with wildcard file check

* Adjust mpiexec configuration

* Add fi statement

* Increase memory allocation

* Change file name in mv command argument

* Increase memory allocation

* Revert to perfdiag file and correct file name

* Remove hardcoded cartopyDataDir path + use updated cartopy shapefiles in evs ve v1.0

---------

+++ scripts/plots/global_ens/exevs_global_ens_atmos_gefs_grid2obs_plots.sh
@@ -41,6 +41,7 @@ while [ $n -le $past_days ] ; do
   day=`$NDATE -$hrs ${VDATE}00|cut -c1-8`
   echo $day
   $USHevs/global_ens/evs_get_gens_atmos_stat_file_link_plots.sh $day "$model_list"
+  export err=$?; err_chk
   n=$((n+1))
 done 
 
@@ -79,8 +80,7 @@ for stats in acc me_mae crpss rmse_spread  ; do
   VARs='PRMSL TMP2m DPT2m UGRD10m VGRD10m RH2m'
   score_types='time_series lead_average'
  else
-  echo $stats is wrong stat
-  exit
+  err_exit "$stats is not a valid metric"
  fi   
 
  for score_type in $score_types ; do
@@ -194,6 +194,7 @@ if [ $run_mpi = yes ] ; then
    mpiexec -np 84 -ppn 84 --cpu-bind verbose,depth cfp ${DATA}/run_all_poe.sh
 else
   ${DATA}/run_all_poe.sh
+  export err=$?; err_chk
 fi
 
 # Cat the plotting log file
@@ -249,15 +250,23 @@ for var in prmsl tmp dpt ugrd vgrd rh; do
                     evs_graphic_level=$level
                 fi
                 if [ $var = prmsl ] ; then
-                    mv lead_average_regional_${domain}_valid_00z_12z_${var}_${stats}.png evs.global_ens.${evs_graphic_stats}.${var}_${evs_graphic_level}.last${past_days}days.fhrmean_valid00z_12z_f384.g003_${domain_new}.png
+                    if [ -f "lead_average_regional_${domain}_valid_00z_12z_${var}_${stats}.png" ]; then
+                        mv lead_average_regional_${domain}_valid_00z_12z_${var}_${stats}.png evs.global_ens.${evs_graphic_stats}.${var}_${evs_graphic_level}.last${past_days}days.fhrmean_valid00z_12z_f384.g003_${domain_new}.png
+                    fi
                 else
-                    mv lead_average_regional_${domain}_valid_00z_12z_${level}${unit}_${var}_${stats}.png evs.global_ens.${evs_graphic_stats}.${var}_${evs_graphic_level}.last${past_days}days.fhrmean_valid00z_12z_f384.g003_${domain_new}.png
+                    if [ -f "lead_average_regional_${domain}_valid_00z_12z_${level}${unit}_${var}_${stats}.png" ]; then
+                        mv lead_average_regional_${domain}_valid_00z_12z_${level}${unit}_${var}_${stats}.png evs.global_ens.${evs_graphic_stats}.${var}_${evs_graphic_level}.last${past_days}days.fhrmean_valid00z_12z_f384.g003_${domain_new}.png
+                    fi
                 fi
                 for lead in 120 240 360; do
                     if [ $var = prmsl ] ; then
-                        mv time_series_regional_${domain}_valid_00z_12z_${var}_${stats}_f${lead}.png evs.global_ens.${evs_graphic_stats}.${var}_${evs_graphic_level}.last${past_days}days.timeseries_valid00z_12z_f${lead}.g003_${domain_new}.png
+                        if [ -f "time_series_regional_${domain}_valid_00z_12z_${var}_${stats}_f${lead}.png" ]; then
+                            mv time_series_regional_${domain}_valid_00z_12z_${var}_${stats}_f${lead}.png evs.global_ens.${evs_graphic_stats}.${var}_${evs_graphic_level}.last${past_days}days.timeseries_valid00z_12z_f${lead}.g003_${domain_new}.png
+                        fi
                     else
-                        mv time_series_regional_${domain}_valid_00z_12z_${level}${unit}_${var}_${stats}_f${lead}.png evs.global_ens.${evs_graphic_stats}.${var}_${evs_graphic_level}.last${past_days}days.timeseries_valid00z_12z_f${lead}.g003_${domain_new}.png
+                        if [ -f "time_series_regional_${domain}_valid_00z_12z_${level}${unit}_${var}_${stats}_f${lead}.png" ]; then
+                            mv time_series_regional_${domain}_valid_00z_12z_${level}${unit}_${var}_${stats}_f${lead}.png evs.global_ens.${evs_graphic_stats}.${var}_${evs_graphic_level}.last${past_days}days.timeseries_valid00z_12z_f${lead}.g003_${domain_new}.png
+                        fi
                     fi
                 done #lead
             done #level
@@ -268,7 +277,7 @@ done     #stats
 tar -cvf evs.plots.${COMPONENT}.${RUN}.${MODELNAME}.${VERIF_CASE}.past${past_days}days.v${VDATE}.tar *.png
 
 if [ $SENDCOM = YES ]; then 
-    cp evs.plots.${COMPONENT}.${RUN}.${MODELNAME}.${VERIF_CASE}.past${past_days}days.v${VDATE}.tar  $COMOUT/.
+    cpreq evs.plots.${COMPONENT}.${RUN}.${MODELNAME}.${VERIF_CASE}.past${past_days}days.v${VDATE}.tar  $COMOUT/.
 fi
 
 if [ $SENDDBN = YES ]; then
